+++ clicommand/agent_start.go
@@ -2,6 +2,7 @@ package clicommand
 
 import (
 	"bufio"
+	"context"
 	"fmt"
 	"io"
 	"net/http"
@@ -89,6 +90,7 @@ type AgentStartConfig struct {
 	HealthCheckAddr            string   `cli:"health-check-addr"`
 	MetricsDatadog             bool     `cli:"metrics-datadog"`
 	MetricsDatadogHost         string   `cli:"metrics-datadog-host"`
+	TracingBackend             string   `cli:"tracing-backend"`
 	Spawn                      int      `cli:"spawn"`
 	LogFormat                  string   `cli:"log-format"`
 	CancelSignal               string   `cli:"cancel-signal"`
@@ -159,6 +161,14 @@ func DefaultConfigFilePaths() (paths []string) {
 	}
 
 	return
+}
+
+// validTracingBackends is a list of valid backends for tracing. This is sanity
+// checked during agent startup so that bootstrapped jobs don't silently have
+// no tracing if an invalid value is given.
+var validTracingBackends = map[string]struct{}{
+	"":        struct{}{},
+	"datadog": struct{}{},
 }
 
 var AgentStartCommand = cli.Command{
@@ -410,6 +420,12 @@ var AgentStartCommand = cli.Command{
 			EnvVar: "BUILDKITE_REDACTED_VARS",
 			Value:  &cli.StringSlice{"*_PASSWORD", "*_SECRET", "*_TOKEN"},
 		},
+		cli.StringFlag{
+			Name:   "tracing-backend",
+			Usage:  "The name of the tracing backend to use.",
+			EnvVar: "BUILDKITE_TRACING_BACKEND",
+			Value:  "",
+		},
 
 		// API Flags
 		AgentRegisterTokenFlag,
@@ -575,6 +591,11 @@ var AgentStartCommand = cli.Command{
 			DatadogHost: cfg.MetricsDatadogHost,
 		})
 
+		// Sanity check supported tracing backends
+		if _, has := validTracingBackends[cfg.TracingBackend]; !has {
+			l.Fatal("The given tracing backend is not supported: %s", cfg.TracingBackend)
+		}
+
 		// AgentConfiguration is the runtime configuration for an agent
 		agentConf := agent.AgentConfiguration{
 			BootstrapScript:            cfg.BootstrapScript,
@@ -601,6 +622,7 @@ var AgentStartCommand = cli.Command{
 			Shell:                      cfg.Shell,
 			RedactedVars:               cfg.RedactedVars,
 			AcquireJob:                 cfg.AcquireJob,
+			TracingBackend:             cfg.TracingBackend,
 		}
 
 		if loader.File != nil {
@@ -838,7 +860,7 @@ func agentShutdownHook(log logger.Logger, cfg AgentStartConfig) {
 
 	// run agent-shutdown hook
 	sh.Promptf("%s", p)
-	if err = sh.RunScript(p, nil); err != nil {
+	if err = sh.RunScript(context.Background(), p, nil); err != nil {
 		log.Error("agent-shutdown hook: %v", err)
 	}
 	w.Close() // goroutine scans until pipe is closed